don't remove the worksqsarn param
even though it isn't actively being used, because we don't want to
delete it's value in case we need it to roll back.

It can, of course, be deleted once we are happy that we won't need to
roll back.

FWIW, this script can be used to get the value back again:

  https://gist.github.com/paulmr/5b430e99ed1f31c4952a35a198a3d3ca