Log performance metrics for android batch notifications
In the metrics dashboard we don't see the android function processing rates. I thought it might
be nice to capture these metrics again for consistency across the two platforms. For android
the function processing rate metric doesn't give us the best indication of performance,
because of the concurrency strategy we're using, but it might be useful to see this data over
time in case the firebase API changes significantly.