feat(report): Remove VPCSecurityGroup from EC2 instances
The `VPCSecurityGroup` group allows ingress and egress to/from 0.0.0.0/0 on all protocols and ports.

The ASG is configured to launch instances in the VPC's private subnets (i.e. not routable from the internet)
and the launch template explicitly sets `AssociatePublicIpAddress` to false.

That is, there is no route from 0.0.0.0/0 to the EC2 instance;
`VPCSecurityGroup` being attached to the EC2 instance has no effect and can be removed.